pyproject.toml
[project]
name = "obsidian"
version = "0.1.0"
description = ""
authors = [
    {name = "iwbc-mzk",email = "46069337+iwbc-mzk@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs-material (>=9.6.4,<10.0.0)",
    "mkdocs-roamlinks-plugin (>=0.3.2,<0.4.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "mkdocs-callouts (>=1.16.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
